requirements.txt
fastapi[all]
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
alembic
pydantic
pydantic-settings
passlib
bcrypt==3.2.2
python-jose[cryptography]
